compositor|refactor: Extract the artifact-resolution schemas
`ResolutionCandidateSchema` and `ArtifactResolutionSchema` now live in `artifact-resolution-schemas.ts`, so the catalog contract and the consistency check shared by both assertions no longer reach into a module named for the plan's dependency graph. Both schemas remain published under the names they had.
