+++ packages/compositor/src/consistency/findResolutionOrderViolations.ts
@@ -1,4 +1,4 @@
-import type { ArtifactResolution } from '../schemas/graph-schemas.ts';
+import type { ArtifactResolution } from '../schemas/artifact-resolution-schemas.ts';
 import type { Violation } from './Violation.ts';
 
 /** One resolution to check, with the path prefix a violation against it is reported under. */
